Move office-event types to its own file
